backend/requirements/base.txt
# Core Django
Django==5.0.1
django-environ==0.11.2

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# REST API
djangorestframework==3.14.0
django-filter==23.5

# CORS
django-cors-headers==4.3.1

# Authentication
djangorestframework-simplejwt==5.3.1

# Async tasks
celery==5.3.6
redis==5.0.1
django-celery-beat==2.6.0

# Validation
pydantic==2.5.3

# Utilities
python-slugify==8.0.1
Pillow==10.2.0

# Security
django-ratelimit==4.1.0
